Added logout test
