+++ src/test/java/com/bitrix/pages/LoginPage.java
@@ -8,7 +8,6 @@ import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
 import org.openqa.selenium.support.PageFactory;
 import org.openqa.selenium.support.ui.ExpectedConditions;
-import org.openqa.selenium.support.ui.Wait;
 import org.openqa.selenium.support.ui.WebDriverWait;
 
 import javax.swing.*;
@@ -26,14 +25,6 @@ public class LoginPage extends TestBase {
 
     @FindBy(css = ".login-btn")
     public WebElement logIn;
-
-    @FindBy(xpath = "//span[@class='menu-popup-item-text']")
-    public WebElement menuPopUp;
-
-    @FindBy(css = "#user-block")
-    public WebElement userBlock;
-
-
 
     public void login(String username, String password) {
         this.username.sendKeys(username);
@@ -71,24 +62,7 @@ public class LoginPage extends TestBase {
         }
     }
 
-    /***
-     * Method navigates to pop up menu item by number ->
-     * 1 - My profile
-     * 2 - Edit profile
-     * 3 - themes
-     * 4 - Configure
-     * 5 - Log out
-     * @param itemNumber - select from 1-5
-     */
 
-    public void selectPopMenu(String itemNumber)
-    {
-        Driver.getDriver().findElement(By.cssSelector("#user-block")).click();
-        WebDriverWait wait = new WebDriverWait(Driver.getDriver(), 10);
-        //Driver.getDriver().manage().window().fullscreen();
-        wait.until(ExpectedConditions.elementToBeClickable(menuPopUp));
-        Driver.getDriver().findElement(By.xpath("(//span[@class='menu-popup-item-text'])["+itemNumber+"]")).click();
-    }
 }
